Add Tile tutorial support
Summary:
Adds tests for all the operations used in the Tile tutorial, plus adds all the necessary support to properly run those tests.

In particular, this implements dilated convolutions and cumsum.

Also adds a file of tutorial tests, and a sandbox. (Not sure whether we should retain the sandbox -- I'm interested in reviewer feedback on that.)

Reviewers: jeremy

Reviewed By: jeremy

Subscribers: jenkins

Differential Revision: https://phb.1026labs.com/D1168