Add OpenCL C code cache
This provides the ability to manually tweak generated kernels by
setting PLAIDML_OPENCL_CACHE to a directory (which must exist).
Initially, generated kernels will be written to this directory
with the filename set to the name of the kernel + ".cl".
Changes to this file will be picked up on the next run.

It is undefined behavior to change the network or any parameters
(such as hardware settings and/or tile scanning) while using the
same PLAIDML_OPENCL_CACHE directory.