Fix flatten with constant variable (#234)
When flattening a contraction that includes a "constant variable" (a variable with a length 1 range), the variable needs to be merged into the constant term of all constraints. Previously, such variables were merely dropped from constraints without changing the constant terms of those constraints. This led to bugs when their value wasn't 0.